Remove uadk_tool dfx --date; __DATE__ and __TIME__ cause problems
Usage of __DATE__ and __TIME__ cause "build not reproducible" warnings
on Debian (-Wdate-time) when building with dpkg-buildpackage, and
since -Werror is set it breaks the build.  Just remove the
functionality; it's not important enough to be worth it.

Signed-off-by: Dan Mick <dan.mick@redhat.com>
(cherry picked from commit 19f650cae960304e3c674992a4c7d5d56a8f4efa)